Rename UniswapSolver to BaselineSolver
