+++ shared/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod amm_pair_provider;
 pub mod arguments;
+pub mod baseline_solver;
 pub mod conversions;
 pub mod current_block;
 pub mod gas_price_estimation;
@@ -10,6 +11,5 @@ pub mod time;
 pub mod token_info;
 pub mod tracing;
 pub mod transport;
-pub mod uniswap_solver;
 
 pub type Web3 = web3::Web3<transport::LoggingTransport<web3::transports::Http>>;
