+++ solver/src/solver.rs
@@ -1,11 +1,11 @@
 use std::{collections::HashSet, fmt::Display, sync::Arc};
 
 use crate::{
+    baseline_solver::BaselineSolver,
     http_solver::{HttpSolver, SolverConfig},
     liquidity::Liquidity,
     naive_solver::NaiveSolver,
     settlement::Settlement,
-    uniswap_solver::UniswapSolver,
 };
 use anyhow::Result;
 use ethcontract::H160;
@@ -39,7 +39,7 @@ pub fn create(
         .into_iter()
         .map(|solver_type| match solver_type {
             SolverType::Naive => Box::new(NaiveSolver {}) as Box<dyn Solver>,
-            SolverType::UniswapBaseline => Box::new(UniswapSolver::new(base_tokens.clone())),
+            SolverType::UniswapBaseline => Box::new(BaselineSolver::new(base_tokens.clone())),
             SolverType::Mip => Box::new(HttpSolver::new(
                 mip_solver_url.clone(),
                 None,
